extract macro create_pool_with_native_on
