+++ cumulus/parachains/integration-tests/emulated/tests/bridges/bridge-hub-westend/src/lib.rs
@@ -30,6 +30,7 @@ mod imports {
 
 	// Cumulus
 	pub use emulated_integration_tests_common::{
+		create_pool_with_native_on,
 		accounts::ALICE,
 		impls::Inspect,
 		test_dry_run_transfer_across_pk_bridge, test_parachain_is_trusted_teleporter,
