+++ cumulus/parachains/integration-tests/emulated/tests/bridges/bridge-hub-westend/src/tests/asset_transfers.rs
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::{create_pool_with_native_on, tests::*};
+use crate::{tests::*};
 use emulated_integration_tests_common::macros::Dmp;
 use xcm::latest::AssetTransferFilter;
